Modifiche

+++ db/migrate/20230818194726_create_grades.rb
@@ -1,7 +1,7 @@
 class CreateGrades < ActiveRecord::Migration[7.0]
   def change
     create_table :grades, id: false, primary_key: [:school_code, :CFstudent, :class_code, :subject_name, :date] do |t|
-      t.integer :value
+      t.float :value
       t.string :CFprof
       t.string :school_code
       t.string :class_code
